sw6.lib imported
Added a modified version of sw6.lib - modified in the sense that I moved the one awkwardly placed package in sw6.girafplace into this.

Merged sw3.server and sw6.server into a single entity. We still need to make some changes to the sw6.girafPlaceClient code in order to remove the last bits of legacy code.

As they say, it *should* be functional :P