Sockets fixed, thead improved.
I managed to miss a significant detail about the server, where sockets weren't properly accepted and read. I've fixed this and improved on the connection code to be less wonky when connections are broken.

Fixed LoggingFormatter to properly prefix all log levels.

Merged SocketThread as a nested class of ConnectionThread. This makes the code a bit more readable and saved me some painful parent referencing.