Contactbook prototype almost finalized.
I've disabled the parser for the purpose of quicker development.

girafscript is intended as a meaningful split of design and function,
but we simply don't have the time or the need to have it ready for
the prototype itself.

I've implemented a lot of hardcoded, bruteforce functionality for the
contactbook, something that is very noticeable on performance. It
now takes several seconds to load a main page (although once it's
loaded, only single app setting modules need to be handled by js
AJAX requests afterwards.)