Contactbook: everything except styling and creating done.
I need some styling done on the answers and message and such.

I haven't implemented replying to a message yet, or creating a new
message.

I'm pretty proud of the js work on the several pages, though.

EVERYTHING needs a cleanup, session support and a bit of logging in stuff.