Implemented image support for new messages. Replies are not supported.
We have unresolved issues with javascript AJAX loading of modules as
well as the text encoding. Sending a UTF-8 encoding header and switching
the MYSQL tables to use that same encoding isn't really helping. Damnit.

Smoothed out a few areas and hardcoded shitloads of GirafImage and
ContactbookMessage functions for retrieval. We have a bit of a mess
to clean up once the final test is over.