Finalizing refactoring.
Model:
nothing

Controller:
Added the 'add' action to the controller. It's an output-less
action (currently) that only reports success/failure back to the
calling client, expected to be a reactive AJAX environment.

View:
Added a jquery plugin, jquery.form, to implement simple AJAX forms.
Touched up the stylesheets.
Moved any function definitions in js out into global scope. Many of
them were inside the jquery ready events and thus their scope was
limited to anything happening within it.
Re-implemented the HTML forms to work the AJAX way. For one, it's
a massive performance boost. For another, it's a cleaner experience.