Issue #144, #145 and #149.
While this build runs there is quite a bit of WIP code lying in
the base. I'm working on an implementation of associative record
connections. Currently it is intended to work much like records
themselves: given a left-right relationship in a table, it is
possible to retrieve any associated keys from one starting point.

The problem right now is that the time and code redundancy reduction
is minimal at best from my current estimates. it *can* be done
even easier.

Perhaps blend the association into GirafRecord as a feature... I
like that.