BROKEN BUILD.
Association *RETRIEVAL* has been implemented, and it seems functional
from base tests.
A counterpart for assocation addition is pending but should be even
simpler to implement now we have the basics covered.

I've rewritten all GirafRecord subclasses to support the new feature
and the code reduction is remarkable.

The current center of attention is the intricate loops involved
when combing through all possible combinations of groups, children,
devices and apps. The number of loops necessary is simply too large
to be properly coded and the overhead is unpleasant to think if.

I'm thinkin JSON controller actions for the views.