Completely eliminated content view dependency.
Contactbook can now, in theory, be displayed without being
injected into main_stub first. It is essentially self-contained
although I've not tested it yet.

We should add a GirafController subclass for app sub controllers,
that forces a form of implementation of getAppId to allow simple
ways of self-refreshing a module.